requirements.txt
google-api-python-client==2.136.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
requests==2.32.3
